// src/app/api/model/thumbnailCreationRequestDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ThumbnailCreationRequestDto { 
    id?: number;
    status?: ThumbnailCreationRequestDto.StatusEnum;
}
export namespace ThumbnailCreationRequestDto {
    export const StatusEnum = {
        Uploading: 'UPLOADING',
        Done: 'DONE',
        Failed: 'FAILED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}